Renamed some components and scss files for better clarity
