+++ client/pages/create-gym-listing.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
 import { useSelector } from 'react-redux';
 import { RootState } from "../redux/rootState";
-import GymForm from "../components/gym-form";
+import AddGymForm from "../components/add-gym-form";
 import Spinner from "../components/spinner";
 
 
-function CreateListing(props: {}) {
+export default function CreateGymListing(props: {}) {
   const [isLoading, setIsLoading] = useState(false);
   const isLoggedIn = useSelector((state: RootState) => state.app.isLoggedIn);
 
@@ -18,11 +18,9 @@ function CreateListing(props: {}) {
   return (
     <>
       <main className="create-main">
-        <GymForm setIsLoading={setIsLoading} />
+        <AddGymForm setIsLoading={setIsLoading} />
       </main>
       <Spinner isLoading={isLoading} />
     </>
   );
 }
-
-export default CreateListing;
