+++ client/pages/edit-gym-listing.tsx
@@ -1,15 +1,15 @@
 import React, { useState } from "react";
-import GymFormEdit from "../components/gym-form-edit";
+import EditGymForm from "../components/edit-gym-form";
 import Spinner from "../components/spinner";
 
-export default function EditListing(props: { gymId: number }) {
+export default function EditGymListing(props: { gymId: number }) {
   const [isLoading, setIsLoading] = useState(false);
   const gymId = props.gymId;
 
   return (
     <>
       <main className="create-main">
-        <GymFormEdit gymId={gymId} setIsLoading={setIsLoading} />
+        <EditGymForm gymId={gymId} setIsLoading={setIsLoading} />
       </main>
       <Spinner isLoading={isLoading} />
     </>
