--- server/public/pages/edit-listing.css
@@ -1,42 +1,0 @@
-.edit-img-container {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-@media (min-width: 720px) {
-  .edit-img-container {
-    flex-direction: row;
-    justify-content: center;
-  }
-}
-.edit-img-container .edit-img-input {
-  margin-bottom: 1rem;
-  margin-left: -2rem;
-}
-@media (min-width: 480px) {
-  .edit-img-container .edit-img-input {
-    margin-left: 0rem;
-  }
-}
-.edit-img-container figure {
-  text-align: center;
-}
-.edit-img-container figure .edit-img-caption {
-  font-weight: bold;
-}
-.edit-img-container figure .edit-img-preview {
-  max-height: 150px;
-}
-@media (min-width: 720px) {
-  .edit-img-container figure .edit-img-preview {
-    max-height: 200px;
-  }
-}
-@media (min-width: 1200px) {
-  .edit-img-container figure .edit-img-preview {
-    max-height: 250px;
-  }
-}
-
-/*# sourceMappingURL=edit-listing.css.map */
