+++ server/public/scss/global/_colors.scss
@@ -2,3 +2,4 @@
 $primary: #D2B85A;
 $secondary: #BFA246;
 $bg-main: #FEFDFB;
+$bg-secondary:  #F5F5F5;
