+++ server/public/scss/pages/auth.scss
@@ -28,7 +28,7 @@
     padding: 2rem;
     border-radius: 15px;
 
-    background-color: #F5F5F5;
+    background-color: $bg-secondary;
     box-shadow: 1px 2px 3px 4px rgba(0, 0, 0, 0.25);
   }
