+++ server/public/scss/pages/create-gym-listing.scss
@@ -13,7 +13,6 @@
   justify-content: center;
   align-items: center;
   padding-top: 1rem;
-  // padding-top: 100px;
 }
 
 .create-form {
