+++ server/public/scss/pages/gym.scss
@@ -7,7 +7,6 @@
 // "xl": 1200px
 
 $primary-box-shadow: 1px 1px 2px 2px rgba(0, 0, 0, 0.25);
-$primary-bg-color: #F5F5F5;
 
 .gym {
   &__main {
@@ -73,7 +72,7 @@ $primary-bg-color: #F5F5F5;
 
     border-radius: 18px;
     box-shadow: $primary-box-shadow;
-    background-color: $primary-bg-color;
+    background-color: $bg-secondary;
 
     @include lg {
       flex-direction: row;
@@ -129,7 +128,7 @@ $primary-bg-color: #F5F5F5;
 
       border-radius: 15px;
       box-shadow: $primary-box-shadow;
-      background-color: $primary-bg-color;
+      background-color: $bg-secondary;
     }
 
     &__inner-container {
@@ -203,7 +202,7 @@ $primary-bg-color: #F5F5F5;
 
       background-color: $primary;
       color: white !important;
-      box-shadow: 1px 1px 2px 2px rgba(0, 0, 0, 0.25);
+      box-shadow: $primary-box-shadow;
       border: solid white 1px;
       border-radius: 5px;
 
@@ -228,7 +227,7 @@ $primary-bg-color: #F5F5F5;
 
       background-color: #ff3333;
       color: white;
-      box-shadow: 1px 1px 2px 2px rgba(0, 0, 0, 0.25);
+      box-shadow: $primary-box-shadow;
       border: solid white 1px;
       border-radius: 5px;
