+++ server/public/scss/styles.scss
@@ -8,8 +8,8 @@
 @use "./components/listing-card.scss";
 @use "./pages/gym.scss";
 @use "./pages/not-found.scss";
-@use "./pages/create-listing.scss";
-@use "./pages/edit-listing.scss";
+@use "./pages/create-gym-listing.scss";
+@use "./pages/edit-gym-listing.scss";
 @use "./pages/auth.scss";
 @use "./pages/account-page.scss";
 @use "./components/reviews.scss";
